Add sub/playground-openapi.yaml
Add reverse engineered OpenAPI specification for the API of the Go
Playground.

I've used the source code at https://cs.opensource.google/go/x/playground/
to discover the API routes and parameters.

So far, the specification is not a perfect match of the real
implementation (the server is quite loose on what it accepts: for
example it doesn't strictly enforce HTTP methods, so often either
GET/POST/PUT/PATCH are accepted), but it should be good enough to get
a way to send data.

Status: tested with https://github.com/openapistack/openapicmd:
  openapi call playground-openapi.yaml -o version
  openapi call playground-openapi.yaml --no-interactive -o play -p 'body=package main;import "fmt";func main(){fmt.Println("OK")}'
  openapi call playground-openapi.yaml --no-interactive -o fmt -p 'body=package main;func main(){fmt.Println("OK")}' -p imports=1
  (however that tool doesn't seem to really map -p values to parameters
defined in the spec, and it doesn't validate them)